==> analyse_oases_global/00_set_paths.R <==
################################################################################
##' @title Set paths for this folder
##' @author Robin Elahi
##' @contact elahi.robin@gmail.com
##' @date 2021-08-24
##' @log 
################################################################################

##### PATHS #####
my_path <- "analyse_oases_global/"
path_to_workspace <- "workspace/"
path_to_figs <- paste(my_path, "analysis_figs/", sep = "")
path_to_output <- paste(my_path, "analysis_output/", sep = "")
path_to_model_rds <- paste(my_path, "model_rds/", sep = "")
path_to_model_output <- paste(my_path,"model_output/model_summaries/", sep = "")
path_to_ignored_model_rds <- "ignore_folder/model_rds/"
